+++ drawer/drawer.hpp
@@ -5,7 +5,7 @@ class Drawer {
     public:
         static void init();
         static bool shouldClose();
-        static void drawGrid(int** grid, int width, int height);
+        static void drawGrid(long long** grid, int width, int height);
         static void close();
 
 };
